+++ app/src/main/java/com/aki/notesapp/navigation/Screens.kt
@@ -3,7 +3,7 @@ package com.aki.notesapp.navigation
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class AddNoteScreenNav(val id: Int?)
+data class AddNoteScreenNav(val id: Long?)
 
 @Serializable
 object NoteScreenNav
